% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{makeIndicator}
\alias{makeIndicator}
\title{Make indicator variables for each levels in a categorical variable.}
\usage{
makeIndicator(x, levels)
}
\arguments{
\item{x}{a character vector of a categorical variable.}

\item{levels}{a character vector of unique levels in \code{x}.}
}
\value{
a data frame of indicator variables of each level in the categorical variable \code{x}.
}
\description{
Make indicator variables for each levels in a categorical variable.
}
\examples{
data(Queensland)
makeIndicator(Queensland[,polActivitiesRaw], c("boycot","union","demonstration","internet","paper","volunteer"))
makeIndicator(Queensland[,"previousVoteRaw"], c("Liberal","Greens","Katter","Labor"))
}

