% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/people_prediction.R
\name{maxWrapper}
\alias{max}
\alias{maxWrapper}
\alias{maximum}
\title{Find an index of a maximum value in a numeric vector.}
\usage{
maxWrapper(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A index of a maximum value in \code{x}
}
\description{
Find an index of a maximum value in a numeric vector.
}
\examples{
maxWrapper(c(1, 3, 2))
maxWrapper(c(0.1, 0.5, 0.7))
}

