% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data_manipulation.R
\docType{data}
\name{QueenslandRaw}
\alias{QueenslandRaw}
\title{Queensland VoteCompass.}
\format{A data frame with 5000 rows and 166 variables:
\describe{
  \item{q1}{How much should Queensland spend to help drought-affected farmers?}
  \item{q2}{Queensland should pay down its debt by leasing public assets.}
  \item{q3}{Teachers' salaries should be tied to student outcomes, rather than seniority.}
  \item{q4}{Religion has no place in state schools.}
  \item{q5}{How much should Queensland do to reduce its greenhouse gas emissions?}
  \item{q6}{Environmental protections are an obstacle to economic growth.}
  \item{q7}{When there is an economic problem, government spending usually makes it worse.}
  \item{q8}{How much of Queensland's public hospital service should be contracted to the private sector?}
  \item{q9}{The government should pay for surgery in a private hospital if wait times in public hospitals are too long.}
  \item{q10}{How many temporary foreign workers should Queensland admit?}
  \item{q11}{How much should the government do to help Aborigines and Torres Strait Islanders find employment?}
  \item{q12}{Indigenous Queenslanders should be able to set their own alcohol laws.}
  \item{q13}{How tough should penalties be for domestic violence be in Queensland?}
  \item{q14}{A member of a bikie gang should receive a tougher sentence than someone else convicted of the same crime.}
  \item{q15}{How strict should Queensland's anti-bikie laws be?}
  \item{q16}{Clubs and bars should not be allowed to sell alcohol after midnight.}
  \item{q17}{How many frontline police officers does Queensland need?}
  \item{q18}{Marijuana should be legal when used for medicinal purposes.}
  \item{q19}{Terminally ill patients should be able to legally end their own lives with medical assistance.}
  \item{q20}{Same-sex couples should have the right to adopt children.}
  \item{q21}{How much should be done to accommodate religious minorities in Queensland?}
  \item{q22}{All immigrants can retain their cultural values without being any less Australian.}
  \item{q23}{How much mining activity should be permitted in the waters around the Great Barrier Reef?}
  \item{q24}{Landowners should be able to prevent mining companies from operating on their property.}
  \item{q25}{How many government workers should Queensland employ?}
  \item{q26}{How much should Queenslanders pay in state taxes?}
  \item{q27}{Rural and regional areas should offer the same government services as cities, even if it means higher taxes for everyone.}
  \item{q28}{Queensland should reduce its debt by cutting services rather than raising taxes.}
  \item{q29}{Traffic congestion in Queensland is best resolved by investing in roads, not in public transport.}
  \item{q30}{How much influence should unions have in Queensland?}
  ...
}}
\source{
\url{http://www.queensland.votecompass.com/}
}
\usage{
QueenslandRaw
}
\description{
A dataset containing the responses to 30 VoteCompass public opinion questions and other demographics of almost 5,000
VoteCompass users.
}
\keyword{datasets}

