% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/people_prediction.R
\name{RCIMeasure}
\alias{RCI}
\alias{RCIMeasure}
\alias{rci}
\title{Calcualte a relative confidence indexn between a winner and other candidates.}
\usage{
RCIMeasure(x, partyNames = NULL)
}
\arguments{
\item{x}{A numerical vector of the winning ratings for each party.}

\item{partyNames}{A character vector of party names in the election race.}
}
\description{
Calcualte a relative confidence indexn between a winner and other candidates.
}
\examples{
RCIMeasure(c(1, 10, 3))
RCIMeasure(c(1, 10, 3), c("Labor","Greens","LibNat"))
}

