% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{recategorize}
\alias{recate}
\alias{recategorize}
\title{Recategorize a character or numerical vector.}
\usage{
recategorize(x, x.new, labels = NA)
}
\arguments{
\item{x}{a character or numerical vector.}

\item{x.new}{a character/numerical vector of the new values for \code{x}. \code{x.new} can also take a form of a matrix or data frame,
where the first column is the original value of \code{x} and the second is the new value of \code{x}.}
}
\value{
a character or numerical vector is recategorized.
}
\description{
Recategorize a character or numerical vector.
}
\examples{
recategorize(c(1,2,5,3,4), c("1" = 1, "2" = 1, "3" = 0))
recategorize(c(1,2,3), cbind(c(1,2,3), c(1,1,0), c(1,-1,0)))
recategorize(c(1,4,2,5,3,NA), data.frame(c(1,2,3), c(1,1,0), c(1, -1, 0)) )
}

