% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/vpl_table.R
\name{vplTable}
\alias{vplTable}
\title{import()}
\usage{
vplTable(x, by = NULL, weight = 1, label.x = NULL, label.by = NULL,
  label.category = NULL, by.subset = NULL, type = c("across", "within",
  "means"), sortData = c("alphabet", "rank", "manual"),
  categoryOrder = NULL)
}
\description{
import()
}

