% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{cleanData}
\alias{clean}
\alias{cleanData}
\alias{rename}
\title{Rename categories of one or more variables in the dataset using an external match file}
\usage{
cleanData(Data, varsClean, matchFileName, ...)
}
\arguments{
\item{Data}{data frame to be processed.}

\item{varsClean}{a character vector of variables to be regrouped.}

\item{matchFileName}{a character string naming a match file.}

\item{...}{other arguments passed on to \code{read.xlsx} for reading the match file.}
}
\value{
a data frame of regrouped variables.
}
\description{
Rename categories of one or more variables in the dataset using an external match file
}
\examples{
cleanData(QueenslandRaw, c("genderRaw","studentRaw","educationRaw"), "~/Desktop/R Packages/QueenslandMatch.xlsx")
cleanData(QueenslandRaw, c("genderRaw","studentRaw","educationRaw"), "QueenslandMatch.xlsx")
}

